+++ local/src/main/java/com/kdjj/local/LocalDataSourceImpl.kt
@@ -1,7 +1,6 @@
-package com.kdjj.local.dataSource
+package com.kdjj.local
 
 import com.kdjj.domain.model.Recipe
-import com.kdjj.local.dataSource.LocalDataSource
 
 class LocalDataSourceImpl: LocalDataSource {
