+++ local/src/main/java/com/kdjj/local/RecipeDatabase.kt
@@ -1,4 +1,4 @@
-package com.kdjj.local.database
+package com.kdjj.local
 
 import android.content.Context
 import androidx.room.Database
